+++ packages/graphql_codegen/lib/builder.dart
@@ -175,20 +175,6 @@ class GraphQLBuilder extends Builder {
       ...Map.fromEntries(
         kGraphQLFileExtensions.map(
           (e) => MapEntry(
-            p.join(_assetsPrefix, '{{dir}}', '{{file}}.${e}'),
-            [
-              p.join(
-                p.relative(config.outputDirectory, from: '/'),
-                '{{dir}}',
-                '{{file}}.${e}.dart',
-              )
-            ],
-          ),
-        ),
-      ),
-      ...Map.fromEntries(
-        kGraphQLFileExtensions.map(
-          (e) => MapEntry(
             p.join(_assetsPrefix, '{{file}}.${e}'),
             [
               p.join(
