+++ packages/graphql_codegen/test/assets/issue_318/graphql/query2/q.graphql
@@ -1,7 +1,3 @@
-type Query {
-    hello: String
-}
-
 query FetchHello {
-    hello
+  hello
 }
