+++ packages/graphql_codegen/test/assets/issue_318/lib/query2/q.graphql.dart
@@ -1,176 +1,5 @@
 import 'package:gql/ast.dart';
 
-enum Enum$__TypeKind {
-  SCALAR,
-  OBJECT,
-  INTERFACE,
-  UNION,
-  ENUM,
-  INPUT_OBJECT,
-  LIST,
-  NON_NULL,
-  $unknown
-}
-
-String toJson$Enum$__TypeKind(Enum$__TypeKind e) {
-  switch (e) {
-    case Enum$__TypeKind.SCALAR:
-      return r'SCALAR';
-    case Enum$__TypeKind.OBJECT:
-      return r'OBJECT';
-    case Enum$__TypeKind.INTERFACE:
-      return r'INTERFACE';
-    case Enum$__TypeKind.UNION:
-      return r'UNION';
-    case Enum$__TypeKind.ENUM:
-      return r'ENUM';
-    case Enum$__TypeKind.INPUT_OBJECT:
-      return r'INPUT_OBJECT';
-    case Enum$__TypeKind.LIST:
-      return r'LIST';
-    case Enum$__TypeKind.NON_NULL:
-      return r'NON_NULL';
-    case Enum$__TypeKind.$unknown:
-      return r'$unknown';
-  }
-}
-
-Enum$__TypeKind fromJson$Enum$__TypeKind(String value) {
-  switch (value) {
-    case r'SCALAR':
-      return Enum$__TypeKind.SCALAR;
-    case r'OBJECT':
-      return Enum$__TypeKind.OBJECT;
-    case r'INTERFACE':
-      return Enum$__TypeKind.INTERFACE;
-    case r'UNION':
-      return Enum$__TypeKind.UNION;
-    case r'ENUM':
-      return Enum$__TypeKind.ENUM;
-    case r'INPUT_OBJECT':
-      return Enum$__TypeKind.INPUT_OBJECT;
-    case r'LIST':
-      return Enum$__TypeKind.LIST;
-    case r'NON_NULL':
-      return Enum$__TypeKind.NON_NULL;
-    default:
-      return Enum$__TypeKind.$unknown;
-  }
-}
-
-enum Enum$__DirectiveLocation {
-  QUERY,
-  MUTATION,
-  SUBSCRIPTION,
-  FIELD,
-  FRAGMENT_DEFINITION,
-  FRAGMENT_SPREAD,
-  INLINE_FRAGMENT,
-  VARIABLE_DEFINITION,
-  SCHEMA,
-  SCALAR,
-  OBJECT,
-  FIELD_DEFINITION,
-  ARGUMENT_DEFINITION,
-  INTERFACE,
-  UNION,
-  ENUM,
-  ENUM_VALUE,
-  INPUT_OBJECT,
-  INPUT_FIELD_DEFINITION,
-  $unknown
-}
-
-String toJson$Enum$__DirectiveLocation(Enum$__DirectiveLocation e) {
-  switch (e) {
-    case Enum$__DirectiveLocation.QUERY:
-      return r'QUERY';
-    case Enum$__DirectiveLocation.MUTATION:
-      return r'MUTATION';
-    case Enum$__DirectiveLocation.SUBSCRIPTION:
-      return r'SUBSCRIPTION';
-    case Enum$__DirectiveLocation.FIELD:
-      return r'FIELD';
-    case Enum$__DirectiveLocation.FRAGMENT_DEFINITION:
-      return r'FRAGMENT_DEFINITION';
-    case Enum$__DirectiveLocation.FRAGMENT_SPREAD:
-      return r'FRAGMENT_SPREAD';
-    case Enum$__DirectiveLocation.INLINE_FRAGMENT:
-      return r'INLINE_FRAGMENT';
-    case Enum$__DirectiveLocation.VARIABLE_DEFINITION:
-      return r'VARIABLE_DEFINITION';
-    case Enum$__DirectiveLocation.SCHEMA:
-      return r'SCHEMA';
-    case Enum$__DirectiveLocation.SCALAR:
-      return r'SCALAR';
-    case Enum$__DirectiveLocation.OBJECT:
-      return r'OBJECT';
-    case Enum$__DirectiveLocation.FIELD_DEFINITION:
-      return r'FIELD_DEFINITION';
-    case Enum$__DirectiveLocation.ARGUMENT_DEFINITION:
-      return r'ARGUMENT_DEFINITION';
-    case Enum$__DirectiveLocation.INTERFACE:
-      return r'INTERFACE';
-    case Enum$__DirectiveLocation.UNION:
-      return r'UNION';
-    case Enum$__DirectiveLocation.ENUM:
-      return r'ENUM';
-    case Enum$__DirectiveLocation.ENUM_VALUE:
-      return r'ENUM_VALUE';
-    case Enum$__DirectiveLocation.INPUT_OBJECT:
-      return r'INPUT_OBJECT';
-    case Enum$__DirectiveLocation.INPUT_FIELD_DEFINITION:
-      return r'INPUT_FIELD_DEFINITION';
-    case Enum$__DirectiveLocation.$unknown:
-      return r'$unknown';
-  }
-}
-
-Enum$__DirectiveLocation fromJson$Enum$__DirectiveLocation(String value) {
-  switch (value) {
-    case r'QUERY':
-      return Enum$__DirectiveLocation.QUERY;
-    case r'MUTATION':
-      return Enum$__DirectiveLocation.MUTATION;
-    case r'SUBSCRIPTION':
-      return Enum$__DirectiveLocation.SUBSCRIPTION;
-    case r'FIELD':
-      return Enum$__DirectiveLocation.FIELD;
-    case r'FRAGMENT_DEFINITION':
-      return Enum$__DirectiveLocation.FRAGMENT_DEFINITION;
-    case r'FRAGMENT_SPREAD':
-      return Enum$__DirectiveLocation.FRAGMENT_SPREAD;
-    case r'INLINE_FRAGMENT':
-      return Enum$__DirectiveLocation.INLINE_FRAGMENT;
-    case r'VARIABLE_DEFINITION':
-      return Enum$__DirectiveLocation.VARIABLE_DEFINITION;
-    case r'SCHEMA':
-      return Enum$__DirectiveLocation.SCHEMA;
-    case r'SCALAR':
-      return Enum$__DirectiveLocation.SCALAR;
-    case r'OBJECT':
-      return Enum$__DirectiveLocation.OBJECT;
-    case r'FIELD_DEFINITION':
-      return Enum$__DirectiveLocation.FIELD_DEFINITION;
-    case r'ARGUMENT_DEFINITION':
-      return Enum$__DirectiveLocation.ARGUMENT_DEFINITION;
-    case r'INTERFACE':
-      return Enum$__DirectiveLocation.INTERFACE;
-    case r'UNION':
-      return Enum$__DirectiveLocation.UNION;
-    case r'ENUM':
-      return Enum$__DirectiveLocation.ENUM;
-    case r'ENUM_VALUE':
-      return Enum$__DirectiveLocation.ENUM_VALUE;
-    case r'INPUT_OBJECT':
-      return Enum$__DirectiveLocation.INPUT_OBJECT;
-    case r'INPUT_FIELD_DEFINITION':
-      return Enum$__DirectiveLocation.INPUT_FIELD_DEFINITION;
-    default:
-      return Enum$__DirectiveLocation.$unknown;
-  }
-}
-
 class Query$FetchHello {
   Query$FetchHello({
     this.hello,
@@ -316,4 +145,3 @@ const documentNodeQueryFetchHello = DocumentNode(definitions: [
     ]),
   ),
 ]);
-const possibleTypesMap = <String, Set<String>>{};
